report-management - refactor toEncryptedBarkReportData
