+++ src/lib/bark/mappers/to-encrypted-bark-report-data.ts
@@ -1,0 +1,17 @@
+import { BarkContext } from "../bark-context";
+import { BarkReportData, EncryptedBarkReportData } from "../bark-models";
+
+export async function toEncryptedBarkReportData(
+  context: BarkContext,
+  reportData: BarkReportData,
+): Promise<EncryptedBarkReportData> {
+  const { textEncryptionService } = context;
+  const { ineligibilityReason, ...otherFields } = reportData;
+  const encryptedIneligibilityReason =
+    await textEncryptionService.getEncryptedData(ineligibilityReason);
+  const encryptedReportData: EncryptedBarkReportData = {
+    ...otherFields,
+    encryptedIneligibilityReason,
+  };
+  return encryptedReportData;
+}
