+++ src/lib/bark/operations/op-edit-report.ts
@@ -5,6 +5,7 @@ import {
 import { CODE } from "@/lib/utilities/bark-code";
 import { BarkContext } from "../bark-context";
 import { updateReport } from "../queries/update-report";
+import { toEncryptedBarkReportData } from "../mappers/to-encrypted-bark-report-data";
 
 export async function opEditReport(
   context: BarkContext,
@@ -12,16 +13,13 @@ export async function opEditReport(
 ): Promise<
   typeof CODE.OK | typeof CODE.ERROR_REPORT_NOT_FOUND | typeof CODE.FAILED
 > {
-  const { dbPool, textEncryptionService } = context;
+  const { dbPool } = context;
   const { reportId, reportData } = args;
-  const { ineligibilityReason, ...otherFields } = reportData;
   try {
-    const encryptedIneligibilityReason =
-      await textEncryptionService.getEncryptedData(ineligibilityReason);
-    const encryptedReportData: EncryptedBarkReportData = {
-      ...otherFields,
-      encryptedIneligibilityReason,
-    };
+    const encryptedReportData = await toEncryptedBarkReportData(
+      context,
+      reportData,
+    );
     const { numUpdated } = await updateReport(dbPool, {
       reportId,
       encryptedReportData,
